added packages and updated unit test
