+++ java-objectcopies/src/main/java/computermodel/ComputerDeep.java
@@ -1,3 +1,5 @@
+package computermodel;
+
 public class ComputerDeep {
 
     private HardDrive hardDrive;
