+++ java-objectcopies/src/main/java/computermodel/HardDrive.java
@@ -1,9 +1,11 @@
-public class Cpu {
+package computermodel;
+
+public class HardDrive {
 
     private String modelNumber;
     private String brandName;
 
-    public Cpu(String modelNumber, String brandName) {
+    public HardDrive(String modelNumber, String brandName) {
         this.modelNumber = modelNumber;
         this.brandName = brandName;
     }
@@ -12,15 +14,8 @@ public class Cpu {
         return modelNumber;
     }
 
-    public void setModelNumber(String modelNumber) {
-        this.modelNumber = modelNumber;
-    }
-
     public String getBrandName() {
         return brandName;
     }
 
-    public void setBrandName(String brandName) {
-        this.brandName = brandName;
-    }
 }
